requirements.txt
fastapi==0.115.4
uvicorn==0.32.0
python-dotenv==1.0.1
redis==5.0.1
SQLAlchemy==2.0.36
scapy==2.5.0
requests==2.32.3
prometheus-client==0.20.0
pytest==8.3.3
joblib==1.5.2
numpy>=1.26.0
PyYAML==6.0.1
scikit-learn==1.5.2
websockets==12.0
flake8==7.1.1
bandit==1.7.9
pytest-asyncio==0.23.8
matplotlib==3.9.0
plotly==5.24.1
pandas==2.2.3
jinja2==3.1.4
reportlab==4.2.5
pillow==11.0.0
aiosmtplib==3.0.2